+++ saleor/graphql/webhook/enums.py
@@ -1,7 +1,7 @@
 import graphene
 
 from ...webhook.event_types import WebhookEventAsyncType, WebhookEventSyncType
-from ..core.descriptions import ADDED_IN_36, ADDED_IN_38, PREVIEW_FEATURE
+from ..core.descriptions import ADDED_IN_36, ADDED_IN_38, ADDED_IN_312, PREVIEW_FEATURE
 from ..core.utils import str_to_enum
 
 checkout_updated_event_enum_description = (
@@ -117,6 +117,12 @@ WEBHOOK_EVENT_DESCRIPTION = {
     WebhookEventAsyncType.PRODUCT_METADATA_UPDATED: (
         "A product metadata is updated." + ADDED_IN_38 + PREVIEW_FEATURE
     ),
+    WebhookEventAsyncType.PRODUCT_MEDIA_CREATED: "A new product media is created."
+    + ADDED_IN_312,
+    WebhookEventAsyncType.PRODUCT_MEDIA_UPDATED: "A product media is updated."
+    + ADDED_IN_312,
+    WebhookEventAsyncType.PRODUCT_MEDIA_DELETED: "A product media is deleted."
+    + ADDED_IN_312,
     WebhookEventAsyncType.PRODUCT_VARIANT_CREATED: "A new product variant is created.",
     WebhookEventAsyncType.PRODUCT_VARIANT_UPDATED: "A product variant is updated.",
     WebhookEventAsyncType.PRODUCT_VARIANT_DELETED: "A product variant is deleted.",
@@ -166,6 +172,7 @@ WEBHOOK_EVENT_DESCRIPTION = {
     ),
     WebhookEventAsyncType.ANY: "All the events.",
     WebhookEventAsyncType.OBSERVABILITY: "An observability event is created.",
+    WebhookEventAsyncType.THUMBNAIL_CREATED: "A thumbnail is created." + ADDED_IN_312,
     WebhookEventSyncType.SHIPPING_LIST_METHODS_FOR_CHECKOUT: (
         "Fetch external shipping methods for checkout."
     ),