+++ saleor/graphql/webhook/tests/mutations/test_webhook_dry_run.py
@@ -238,6 +238,9 @@ def async_subscription_webhooks_with_root_objects(
     subscription_product_updated_webhook,
     subscription_product_created_webhook,
     subscription_product_deleted_webhook,
+    subscription_product_media_updated_webhook,
+    subscription_product_media_created_webhook,
+    subscription_product_media_deleted_webhook,
     subscription_product_metadata_updated_webhook,
     subscription_product_variant_created_webhook,
     subscription_product_variant_updated_webhook,
@@ -326,6 +329,7 @@ def async_subscription_webhooks_with_root_objects(
     warehouse,
     translated_attribute,
     transaction_item,
+    product_media_image,
 ):
     events = WebhookEventAsyncType
     attr = numeric_attribute
@@ -401,6 +405,18 @@ def async_subscription_webhooks_with_root_objects(
         events.PRODUCT_CREATED: [subscription_product_created_webhook, product],
         events.PRODUCT_UPDATED: [subscription_product_updated_webhook, product],
         events.PRODUCT_DELETED: [subscription_product_deleted_webhook, product],
+        events.PRODUCT_MEDIA_CREATED: [
+            subscription_product_media_created_webhook,
+            product_media_image,
+        ],
+        events.PRODUCT_MEDIA_UPDATED: [
+            subscription_product_media_updated_webhook,
+            product_media_image,
+        ],
+        events.PRODUCT_MEDIA_DELETED: [
+            subscription_product_media_deleted_webhook,
+            product_media_image,
+        ],
         events.PRODUCT_METADATA_UPDATED: [
             subscription_product_metadata_updated_webhook,
             product,