+++ saleor/plugins/base_plugin.py
@@ -33,6 +33,7 @@ from ..payment.interface import (
     PaymentGateway,
     TransactionActionData,
 )
+from ..thumbnail.models import Thumbnail
 from .models import PluginConfiguration
 
 if TYPE_CHECKING:
@@ -57,6 +58,7 @@ if TYPE_CHECKING:
         Category,
         Collection,
         Product,
+        ProductMedia,
         ProductType,
         ProductVariant,
     )
@@ -792,6 +794,24 @@ class BasePlugin:
     # updated.
     product_updated: Callable[["Product", Any], Any]
 
+    # Trigger when product media is created.
+    #
+    # Overwrite this method if you need to trigger specific logic after a product media
+    # is created.
+    product_media_created: Callable[["ProductMedia", Any], Any]
+
+    # Trigger when product media is updated.
+    #
+    # Overwrite this method if you need to trigger specific logic after a product media
+    # is updated.
+    product_media_updated: Callable[["ProductMedia", Any], Any]
+
+    # Trigger when product media is created.
+    #
+    # Overwrite this method if you need to trigger specific logic after a product media
+    # is deleted.
+    product_media_deleted: Callable[["ProductMedia", Any], Any]
+
     # Trigger when product metadata is updated.
     #
     # Overwrite this method if you need to trigger specific logic after a product
@@ -908,6 +928,9 @@ class BasePlugin:
     # Overwrite this method if you need to trigger specific logic after a staff user is
     # deleted.
     staff_deleted: Callable[["User", Any], Any]
+
+    # Trigger when thumbnail is updated.
+    thumbnail_created: Callable[["Thumbnail", Any], Any]
 
     # Trigger when tracking number is updated.
     tracking_number_updated: Callable[["Fulfillment", Any], Any]