+++ saleor/webhook/payloads.py
@@ -34,10 +34,11 @@ from ..page.models import Page
 from ..payment import ChargeStatus
 from ..plugins.webhook.utils import from_payment_app_id
 from ..product import ProductMediaTypes
-from ..product.models import Collection, Product
+from ..product.models import Collection, Product, ProductMedia
 from ..shipping.interface import ShippingMethodData
 from ..tax.models import TaxClassCountryRate
 from ..tax.utils import get_charge_taxes_for_order
+from ..thumbnail.models import Thumbnail
 from ..warehouse.models import Stock, Warehouse
 from . import traced_payload_generator
 from .event_types import WebhookEventAsyncType
@@ -1473,3 +1474,15 @@ def generate_transaction_action_request_payload(
         "meta": generate_meta(requestor_data=generate_requestor(requestor)),
     }
     return json.dumps(payload, cls=CustomJsonEncoder)
+
+
+@traced_payload_generator
+def generate_thumbnail_payload(thumbnail: Thumbnail):
+    thumbnail_id = graphene.Node.to_global_id("Thumbnail", thumbnail.id)
+    return json.dumps({"id": thumbnail_id})
+
+
+@traced_payload_generator
+def generate_product_media_payload(media: ProductMedia):
+    product_media_id = graphene.Node.to_global_id("ProductMedia", media.id)
+    return json.dumps({"id": product_media_id})